requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.13
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
pillow==10.3.0
prompt-toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.5
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
